ganymede/x/ganymede/types/messages_users_test.go
package types

import (
	"testing"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
	"github.com/jack139/ganymede/ganymede/testutil/sample"
)

func TestMsgCreateUsers_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgCreateUsers
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgCreateUsers{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgCreateUsers{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}

func TestMsgUpdateUsers_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgUpdateUsers
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgUpdateUsers{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgUpdateUsers{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}

func TestMsgDeleteUsers_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgDeleteUsers
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgDeleteUsers{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgDeleteUsers{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
